requirements.txt
chromadb>=0.5.0
numpy>=1.24
pillow>=10.0
pymupdf>=1.23.0
pytesseract>=0.3.10
pyyaml>=6.0.1
tqdm>=4.66
transformers>=4.45.0
accelerate>=0.34.0
sentence-transformers>=3.0.1
torch>=2.1.0
